Split main file up by class
Helps to make reading through each portion much easier. Main file was getting too long to modify easily.
